+++ app/charts/column/columns-grouped-state-props.ts
@@ -53,7 +53,7 @@ export const useColumnsGroupedStateVariables = (
     dimensionsByIri,
     observations,
   });
-  const numericalYVariables = useNumericalYVariables(y, {
+  const numericalYVariables = useNumericalYVariables("column", y, {
     measuresByIri,
   });
   const numericalYErrorVariables = useNumericalYErrorVariables(y, {