+++ app/components/react-grid.tsx
@@ -27,12 +27,12 @@ export const availableHandles: ResizeHandle[] = [
 /** In grid unit */
 const MAX_H = 10;
 const INITIAL_H = 7;
-const MIN_H = 5;
+export const MIN_H = 5;
 
 /** In grid unit */
-const MAX_W = 4;
+export const MAX_W = 4;
 
-const COLS = { lg: 4, md: 2, sm: 1, xs: 1, xxs: 1 };
+export const COLS = { lg: 4, md: 2, sm: 1, xs: 1, xxs: 1 };
 const ROW_HEIGHT = 100;
 
 const useStyles = makeStyles((theme: Theme) => ({